Add support for the Æther
Some people might not want to be tracked _into_ a space (say a bedroom), but might still want things to react when they enter common space. Enter the Æther. Rooms can be specified as "ætherial", which will only track changes from the other room, but not publish information about the room